implementa as exceptions nos controller e service de usuario

+++ src/services/UserServices.ts
@@ -1,8 +1,9 @@
 import { PostgresDataSource } from "../database/data-source";
 import UserAlreadyExistException from "../exceptions/userExceptions";
 import { User } from "../database/entity/User";
+import NullDataException from "../exceptions/serveExceptions";
 
-const userRepository =  PostgresDataSource.getRepository(User);
+const userRepository = PostgresDataSource.getRepository(User);
 
 class UserServices {
   Create = async (user: User) => {
@@ -12,6 +13,11 @@ class UserServices {
       })
     ) {
       throw new UserAlreadyExistException("User already exists");
+    }
+    if (user == null) {
+      throw new NullDataException(
+        "There are parameters that were not declared correctly"
+      );
     }
     const userCreated = await userRepository.create(user);
     const userSaved = await userRepository.save(userCreated);
@@ -23,30 +29,44 @@ class UserServices {
     return users;
   };
 
-  Find = async (userId:string ) =>{
+  Find = async (userId: string) => {
+    if (userId == null) {
+      throw new NullDataException(
+        "There are parameters that were not declared correctly"
+      );
+    }
     const user = await userRepository.find({
-      where:{
-        id:userId
+      where: {
+        id: userId,
       },
-      relations:{
-        tasks:true
-      }
+      relations: {
+        tasks: true,
+      },
     });
     return user;
   };
 
-  Update = async(userId:string,content) =>{
-      const user = await userRepository.findOneBy({id:userId});
-      const merge = await userRepository.merge(user,user.name=content);
-      const result = await userRepository.save(merge);
-      return result;
+  Update = async (userId: string, content) => {
+    if (userId == null && content == null) {
+      throw new NullDataException(
+        "There are parameters that were not declared correctly"
+      );
+    }
+    const user = await userRepository.findOneBy({ id: userId });
+    const merge = await userRepository.merge(user, (user.name = content));
+    const result = await userRepository.save(merge);
+    return result;
   };
 
-  Delete =async (userId:string) => {
-      const result = await userRepository.delete(userId)
-      return result;
+  Delete = async (userId: string) => {
+    if (userId == null) {
+      throw new NullDataException(
+        "There are parameters that were not declared correctly"
+      );
+    }
+    const result = await userRepository.delete(userId);
+    return result;
   };
+}
 
-  };
-
-export const userServices = new UserServices()
+export const userServices = new UserServices();
